finished user testing the app needs some unit tests though
