+++ src/main/java/de/hitec/nhplus/Main.java
@@ -1,11 +1,12 @@
 package de.hitec.nhplus;
 
-import de.hitec.nhplus.controller.MainWindowController;
+import de.hitec.nhplus.controller.LoginController;
 import de.hitec.nhplus.datastorage.ConnectionBuilder;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
@@ -24,7 +25,46 @@ public class Main extends Application {
         this.primaryStage = primaryStage;
 
         // Start with login window
-        new MainWindowController().showLoginWindow();
+        this.showLoginWindow();
+    }
+
+        /**
+     * Show the login window first
+     */
+    public void showLoginWindow() {
+        try {
+            // Load the login FXML
+            FXMLLoader loader = new FXMLLoader();
+            loader.setLocation(Main.class.getResource("/de/hitec/nhplus/AuthView.fxml"));
+            AnchorPane loginPane = loader.load();
+
+            // Get the controller and set reference to main app
+            LoginController loginController = loader.getController();
+            loginController.setMainApp(this);  // Now this will work!
+
+            // Create login scene
+            Scene loginScene = new Scene(loginPane);
+            
+            // Configure login stage
+            Stage loginStage = new Stage();
+            loginStage.setTitle("NHPlus - Anmeldung");
+            loginStage.setScene(loginScene);
+            loginStage.setResizable(false);
+            loginStage.centerOnScreen();
+            
+            // Handle close request
+            loginStage.setOnCloseRequest(event -> {
+                ConnectionBuilder.closeConnection();
+                Platform.exit();
+                System.exit(0);
+            });
+            
+            loginStage.show();
+
+        } catch (IOException exception) {
+            System.err.println("Error loading login window: " + exception.getMessage());
+            exception.printStackTrace();
+        }
     }
 
     /**
