+++ src/main/java/module-info.java
@@ -4,6 +4,7 @@ module de.hitec.nhplus {
 
     requires org.controlsfx.controls;
     requires java.sql;
+    requires javafx.graphics;
     requires bcrypt;
     requires org.xerial.sqlitejdbc;
     opens de.hitec.nhplus to javafx.fxml;
