+++ src/main/resources/de/hitec/nhplus/MainWindowView.fxml
@@ -5,6 +5,26 @@
 <?import javafx.scene.layout.*?>
 
 <BorderPane fx:id="mainBorderPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="688.0" prefWidth="926.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.hitec.nhplus.controller.MainWindowController" stylesheets="@Application.css">
+   
+   <!-- Top Menu Bar -->
+   <top>
+      <MenuBar>
+         <menus>
+            <Menu mnemonicParsing="false" text="Datei">
+               <items>
+                  <MenuItem mnemonicParsing="false" onAction="#handleLogout" text="Abmelden" />
+               </items>
+            </Menu>
+            <Menu mnemonicParsing="false" text="Ansicht">
+               <items>
+                  <MenuItem mnemonicParsing="false" onAction="#handleShowAllPatient" text="Patienten" />
+                  <MenuItem mnemonicParsing="false" onAction="#handleShowAllTreatments" text="Behandlungen" />
+               </items>
+            </Menu>
+         </menus>
+      </MenuBar>
+   </top>
+   
    <left>
       <VBox id="vBox" alignment="CENTER" spacing="50.0" styleClass="vBox" stylesheets="@Application.css" BorderPane.alignment="CENTER">
          <children>
@@ -14,11 +34,14 @@
                </VBox.margin>
                <opaqueInsets>
                   <Insets />
-               </opaqueInsets></Button>
+               </opaqueInsets>
+            </Button>
+            
             <Button alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#handleShowAllTreatments" prefWidth="105.0" text="Behandlungen">
                <VBox.margin>
                   <Insets bottom="50.0" left="10.0" right="10.0" top="50.0" />
-               </VBox.margin></Button>
+               </VBox.margin>
+            </Button>
          </children>
       </VBox>
    </left>
